% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet.R
\name{greet}
\alias{greet}
\title{Greet a user}
\usage{
greet(name = "User")
}
\arguments{
\item{name}{character; name of user}
}
\value{
nothing
}
\description{
Prints a greeting for a user.
}
\examples{
greet()
greet("Sophie")
}
